+++ plugins/python/src/main/kotlin/io/paddle/plugin/python/dependencies/index/PyPackagesRepositoryIndexer.kt
@@ -4,6 +4,7 @@ import io.ktor.client.request.*
 import io.ktor.client.statement.*
 import io.paddle.plugin.python.dependencies.index.distributions.PyDistributionInfo
 import io.paddle.plugin.python.dependencies.index.metadata.JsonPackageMetadataInfo
+import io.paddle.plugin.python.dependencies.index.utils.*
 import kotlinx.serialization.decodeFromString
 import org.jsoup.Jsoup
