--- plugins/python/src/main/kotlin/io/paddle/plugin/python/dependencies/index/Utils.kt
@@ -1,55 +1,0 @@
-package io.paddle.plugin.python.dependencies.index
-
-import io.ktor.client.*
-import io.ktor.client.engine.cio.*
-import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.json.Json
-
-private const val THREADS_COUNT = 24
-private const val TIMEOUT_MS = 5000L
-
-typealias PyPackageName = String
-typealias PyPackagesRepositoryUrl = String
-
-internal val jsonParser = Json {
-    ignoreUnknownKeys = true
-}
-
-internal val httpClient = HttpClient(CIO) {
-    followRedirects = true
-    expectSuccess = false
-    engine {
-        threadsCount = THREADS_COUNT
-        maxConnectionsCount = 1000
-        endpoint {
-            connectAttempts = 5
-            connectTimeout = TIMEOUT_MS
-            requestTimeout = TIMEOUT_MS
-            socketTimeout = TIMEOUT_MS
-        }
-    }
-}
-
-fun PyPackagesRepositoryUrl.join(urlPart: String): String = "${this.trimEnd('/')}/${urlPart.trimEnd('/')}/"
-
-internal fun Iterable<String>.letters(): Set<Char> = flatMap { it.toSet() }.toSet()
-
-internal fun <T> Iterable<T>.withIndexAt(start: Int) = withIndex().map { IndexedValue(it.index + start, it.value) }
-
-internal fun ByteArray.compare(other: ByteArray) = this.compare(0, size, other)
-
-internal fun ByteArray.compare(start: Int, size: Int, other: ByteArray): Int {
-    if (size != other.size) return size - other.size
-    //compare elements values
-    for (i in 0 until size) {
-        val diff = this[i + start] - other[i]
-
-        if (diff != 0) return diff
-    }
-    return 0
-}
-
-private val emptyByteArray = ByteArray(0)
-internal fun emptyByteArray() = emptyByteArray
-
-class WrappedSerialDescriptor(override val serialName: String, original: SerialDescriptor) : SerialDescriptor by original
