+++ plugins/python/src/main/kotlin/io/paddle/plugin/python/dependencies/index/wordlist/PackedWordList.kt
@@ -1,6 +1,6 @@
 package io.paddle.plugin.python.dependencies.index.wordlist
 
-import io.paddle.plugin.python.dependencies.index.*
+import io.paddle.plugin.python.dependencies.index.utils.*
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.builtins.SetSerializer
 import kotlinx.serialization.builtins.serializer
